backend/requirements.txt
# ═══════════════════════════════════════════════════════════════
# APIS v5.0 — Backend Dependencies
# ═══════════════════════════════════════════════════════════════

# --- Core Framework ---
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
python-dotenv==1.0.1

# --- Database ---
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
geoalchemy2==0.17.1
alembic==1.14.1

# --- Redis ---
redis==5.2.1

# --- Celery (Async Tasks) ---
celery[redis]==5.4.0

# --- Geospatial ---
shapely==2.0.6
pyproj==3.7.1
rasterio==1.4.3
overpy==0.7

# --- Satellite Data ---
sentinelsat==1.3.0

# --- Computer Vision & ML ---
ultralytics==8.3.52
opencv-python-headless==4.10.0.84
scikit-image==0.24.0
torch==2.5.1
torchvision==0.20.1
numpy==1.26.4
pillow==11.1.0

# --- AI / LLM ---
google-generativeai==0.8.4

# --- PDF Generation ---
reportlab==4.2.5

# --- Selenium (PG Portal) ---
undetected-chromedriver==3.5.5
selenium==4.27.1
2captcha-python==1.5.0

# --- Twilio (WhatsApp/SMS) ---
twilio==9.4.4

# --- HTTP ---
httpx==0.28.1
requests==2.32.3

# --- Weather ---
# (uses httpx — no extra dependency)

# --- Cloud / Storage ---
boto3==1.35.81

# --- Monitoring ---
mlflow==2.19.0

# --- Airflow (DAG definitions only — runtime provided by MWAA) ---
apache-airflow==2.10.4

# --- Utilities ---
python-json-logger==3.2.1
sqlparse==0.5.3
